+++ Libplanet/Net/Protocols/RoutingTable.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Linq;
 using Serilog;
 
@@ -12,6 +13,7 @@ namespace Libplanet.Net.Protocols
     {
         private readonly Address _address;
         private readonly KBucket[] _buckets;
+        private readonly PeerState[] _staticPeerStates;
 
         private readonly ILogger _logger;
 
@@ -21,13 +23,17 @@ namespace Libplanet.Net.Protocols
         /// <param name="address"><see cref="Address"/> of this peer.</param>
         /// <param name="tableSize">The number of buckets in the table.</param>
         /// <param name="bucketSize">The size of a single bucket.</param>
+        /// <param name="staticPeers">
+        /// The list of <see cref="Peer"/>s to keep in routing table permanently.
+        /// The <see cref="Peer"/>s in the list won't be removed in any cases.</param>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown when <paramref name="tableSize"/> or <paramref name="bucketSize"/> is
         /// less then or equal to 0.</exception>
         public RoutingTable(
             Address address,
             int tableSize = Kademlia.TableSize,
-            int bucketSize = Kademlia.BucketSize)
+            int bucketSize = Kademlia.BucketSize,
+            IImmutableSet<BoundPeer> staticPeers = null)
         {
             if (tableSize <= 0)
             {
@@ -50,6 +56,11 @@ namespace Libplanet.Net.Protocols
             {
                 _buckets[i] = new KBucket(BucketSize, random, _logger);
             }
+
+            var initialized = DateTimeOffset.UtcNow;
+            _staticPeerStates = staticPeers is null
+                ? new PeerState[0]
+                : staticPeers.Select(boundPeer => new PeerState(boundPeer, initialized)).ToArray();
         }
 
         /// <summary>
@@ -65,22 +76,40 @@ namespace Libplanet.Net.Protocols
         /// <summary>
         /// The number of peers in the table.
         /// </summary>
-        public int Count => _buckets.Sum(bucket => bucket.Count);
+        public int Count => _buckets.Sum(bucket => bucket.Count) + _staticPeerStates.Length;
 
         /// <summary>
-        /// An <see cref="IEnumerable{T}"/> of peers in the table.
+        /// An <see cref="IReadOnlyList{T}"/> of static peers in the table.
         /// </summary>
-        public IEnumerable<BoundPeer> Peers => NonEmptyBuckets
-            .SelectMany((bucket, _) => bucket.Peers)
-            .ToList();
+        public IReadOnlyList<BoundPeer> StaticPeers => _staticPeerStates
+            .Select(state => state.Peer)
+            .ToArray();
 
         /// <summary>
-        /// An <see cref="IEnumerable{T}"/> of <see cref="PeerState"/> of peers in the table.
+        /// An <see cref="IReadOnlyList{T}"/> of peers in the table.
         /// </summary>
-        public IEnumerable<PeerState> PeerStates =>
-            NonEmptyBuckets.SelectMany(bucket => bucket.PeerStates);
+        public IReadOnlyList<BoundPeer> Peers => StaticPeers
+            .Union(NonEmptyBuckets.SelectMany(bucket => bucket.Peers))
+            .ToArray();
 
-        internal IEnumerable<IEnumerable<BoundPeer>> CachesToCheck
+        /// <summary>
+        /// An <see cref="IReadOnlyList{T}"/> of <see cref="PeerState"/> of
+        /// static peers in the table.
+        /// </summary>
+        public IReadOnlyList<PeerState> StaticPeerStates => _staticPeerStates.ToArray();
+
+        /// <summary>
+        /// An <see cref="IReadOnlyList{T}"/> of <see cref="PeerState"/> of peers in the table.
+        /// </summary>
+        public IReadOnlyList<PeerState> PeerStates => StaticPeerStates
+            .Union(NonEmptyBuckets.SelectMany(bucket => bucket.PeerStates))
+            .ToArray();
+
+        // For unit tests
+        internal IReadOnlyList<BoundPeer> NonStaticPeers =>
+            NonEmptyBuckets.SelectMany(bucket => bucket.Peers).ToArray();
+
+        internal IReadOnlyList<IReadOnlyList<BoundPeer>> CachesToCheck
         {
             get
             {
@@ -88,23 +117,24 @@ namespace Libplanet.Net.Protocols
                     bucket => bucket.ReplacementCache
                         .OrderBy(kv => kv.Value)
                         .Select(kv => kv.Key)
-                );
+                        .ToArray()
+                ).ToArray();
             }
         }
 
-        internal IEnumerable<KBucket> NonFullBuckets
+        internal IReadOnlyList<KBucket> NonFullBuckets
         {
             get
             {
-                return _buckets.Where(bucket => !bucket.IsFull());
+                return _buckets.Where(bucket => !bucket.IsFull()).ToArray();
             }
         }
 
-        internal IEnumerable<KBucket> NonEmptyBuckets
+        internal IReadOnlyList<KBucket> NonEmptyBuckets
         {
             get
             {
-                return _buckets.Where(bucket => !bucket.IsEmpty());
+                return _buckets.Where(bucket => !bucket.IsEmpty()).ToArray();
             }
         }
 
@@ -145,7 +175,8 @@ namespace Libplanet.Net.Protocols
         /// an element with the specified key; otherwise, <see langword="false"/>.</returns>
         public bool Contains(BoundPeer peer)
         {
-            return BucketOf(peer).Contains(peer);
+            return BucketOf(peer).Contains(peer) ||
+                   _staticPeerStates.Any(state => state.Peer.Address.Equals(peer.Address));
         }
 
         /// <summary>
@@ -156,13 +187,10 @@ namespace Libplanet.Net.Protocols
         /// <returns>A <see cref="BoundPeer"/> whose <see cref="Address"/> matches
         /// the given <paramref name="addr"/>.</returns>
         public BoundPeer GetPeer(Address addr) =>
-            _buckets
-                .Where(b => !b.IsEmpty())
-                .SelectMany(b => b.Peers)
-                .FirstOrDefault(peer => peer.Address.Equals(addr));
+            Peers.FirstOrDefault(peer => peer.Address.Equals(addr));
 
         /// <summary>
-        /// Removes all peers in the table.
+        /// Removes all peers in the table. This method does not affect static peers.
         /// </summary>
         public void Clear()
         {
@@ -181,7 +209,7 @@ namespace Libplanet.Net.Protocols
         /// <param name="includeTarget">A boolean value indicates to include a peer with
         /// <see cref="Address"/> of <paramref name="target"/> in return value or not.</param>
         /// <returns>An enumerable of <see cref="BoundPeer"/>.</returns>
-        public IEnumerable<BoundPeer> Neighbors(Peer target, int k, bool includeTarget)
+        public IReadOnlyList<BoundPeer> Neighbors(Peer target, int k, bool includeTarget)
         {
             return Neighbors(target.Address, k, includeTarget);
         }
@@ -197,8 +225,9 @@ namespace Libplanet.Net.Protocols
         /// <param name="includeTarget">A boolean value indicates to include a peer with
         /// <see cref="Address"/> of <paramref name="target"/> in return value or not.</param>
         /// <returns>An enumerable of <see cref="BoundPeer"/>.</returns>
-        public IEnumerable<BoundPeer> Neighbors(Address target, int k, bool includeTarget)
+        public IReadOnlyList<BoundPeer> Neighbors(Address target, int k, bool includeTarget)
         {
+            // TODO: Should include static peers?
             var sorted = _buckets
                 .Where(b => !b.IsEmpty())
                 .SelectMany(b => b.Peers)
@@ -214,7 +243,7 @@ namespace Libplanet.Net.Protocols
                 ? sorted
                 : sorted.Where(peer => !peer.Address.Equals(target));
 
-            return peers.Take(maxCount);
+            return peers.Take(maxCount).ToArray();
         }
 
         /// <summary>
@@ -232,7 +261,17 @@ namespace Libplanet.Net.Protocols
                 throw new ArgumentNullException(nameof(peer));
             }
 
-            BucketOf(peer).Check(peer, start, end);
+            PeerState staticState =
+                _staticPeerStates.FirstOrDefault(state => state.Peer.Address.Equals(peer.Address));
+            if (staticState is null)
+            {
+                BucketOf(peer).Check(peer, start, end);
+            }
+            else
+            {
+                staticState.LastChecked = start;
+                staticState.Latency = end - start;
+            }
         }
 
         internal void AddPeer(BoundPeer peer, DateTimeOffset updated)
@@ -250,19 +289,51 @@ namespace Libplanet.Net.Protocols
                 );
             }
 
-            _logger.Debug(
-                "Adding a peer {Peer} to peer {Address}'s routing table...",
-                peer,
-                _address);
-            BucketOf(peer).AddPeer(peer, updated);
+            PeerState staticState =
+                _staticPeerStates.FirstOrDefault(state => state.Peer.Address.Equals(peer.Address));
+            if (staticState is null)
+            {
+                _logger.Debug(
+                    "Adding a peer {Peer} to peer {Address}'s routing table...",
+                    peer,
+                    _address);
+                BucketOf(peer).AddPeer(peer, updated);
+            }
+            else
+            {
+                _logger.Verbose("The peer {Peer} to be added is already in static peer.", peer);
+
+                // This done because peer's other attribute except public key might be changed.
+                // (eg. public IP address, endpoint)
+                staticState.Peer = peer;
+                staticState.LastUpdated = updated;
+            }
         }
 
-        internal IEnumerable<BoundPeer> PeersToBroadcast(Address? except, int min = 10)
+        internal IReadOnlyList<BoundPeer> PeersToBroadcast(Address? except, int min = 10)
         {
             List<BoundPeer> peers = NonEmptyBuckets
                 .Select(bucket => bucket.GetRandomPeer(except))
                 .Where(peer => !(peer is null)).ToList();
-            var count = peers.Count;
+            int count;
+            if (_staticPeerStates.Any())
+            {
+                var random = new Random();
+                var excepted = _staticPeerStates.Where(
+                    state => except is null || !state.Peer.Address.Equals(except));
+                var selected = _staticPeerStates[random.Next(_staticPeerStates.Length)];
+                peers.Add(selected.Peer);
+
+                count = peers.Count;
+                if (count < min)
+                {
+                    peers.AddRange(
+                        excepted.Where(state => peers.Contains(state.Peer)).Take(min - peers.Count)
+                            .Select(state => state.Peer));
+                }
+            }
+
+            count = peers.Count;
             if (count < min)
             {
                 peers.AddRange(Peers
@@ -275,12 +346,14 @@ namespace Libplanet.Net.Protocols
             return peers;
         }
 
-        internal IEnumerable<BoundPeer> PeersToRefresh(TimeSpan maxAge)
-        {
-            return NonEmptyBuckets
-                .Where(bucket => bucket.Tail.LastUpdated + maxAge < DateTimeOffset.UtcNow)
-                .Select(bucket => bucket.Tail.Peer);
-        }
+        internal IReadOnlyList<BoundPeer> PeersToRefresh(TimeSpan maxAge) => NonEmptyBuckets
+            .Where(bucket => bucket.Tail.LastUpdated + maxAge < DateTimeOffset.UtcNow)
+            .Select(bucket => bucket.Tail.Peer)
+            .Union(
+                _staticPeerStates
+                    .Where(state => state.LastUpdated + maxAge < DateTimeOffset.UtcNow)
+                    .Select(state => state.Peer).Take(1))
+            .ToArray();
 
         internal bool RemoveCache(BoundPeer peer)
         {
@@ -291,7 +364,7 @@ namespace Libplanet.Net.Protocols
         internal KBucket BucketOf(BoundPeer peer)
         {
             int index = GetBucketIndexOf(peer.Address);
-            return _buckets[index];
+            return BucketOf(index);
         }
 
         internal KBucket BucketOf(int level)